solana-allocator: move custom Solana allocator to its own crate
Having the allocator in separate crate helps with encapsulation and
reasoning about APIs and boundaries between solana-ibc and the
allocator.
